requirements.txt
pandas==1.3.5
Pillow==9.4.0
plotly==5.11.0
streamlit==1.16.0
